\name{dataonelibs-package}
\alias{dataonelibs-package}
\docType{package}
\title{
DataONE R Client Libraries
}
\description{
A package that provides a set of Java libraries that are dependencies for 
    the DataONE R client to work properly.  It is split into a separate R library to 
    make updates to the DataONE R Client package easier.
}
\details{
\tabular{ll}{
Package: \tab dataonelibs\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-02-25\cr
SystemRequirements: \tab Java (>= 1.6)\cr
License: \tab file LICENSE\cr
Packaged: \tab 2013-02-25 20:54:27 UTC; jones\cr
Repository: \tab CRAN\cr
Date/Publication: \tab 2013-02-25 20:49:45\cr
URL: \tab http://www.dataone.org/investigator-toolkit/dataone_r\cr
BugReports: \tab https://redmine.dataone.org\cr
Built: \tab R 2.15.1; ; 2013-02-25 21:42:52 UTC; unix\cr
}
}
\author{
Matthew Jones, Rob Nahf
}
\references{
The DataONE API that is implemented in the enclosed libraries can be found at https://mule1.dataone.org/ArchitectureDocs-current/
}